% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_stats.R
\name{cran_stats}
\alias{cran_stats}
\title{cran_stats}
\usage{
cran_stats(packages, use_cache = TRUE)
}
\arguments{
\item{packages}{packages}

\item{use_cache}{logical, should cached data be used? Default: TRUE. If set to FALSE, it will
re-query download stats and update cache.}
}
\value{
data.frame
}
\description{
monthly download stats of cran package(s)
}
\examples{
\dontrun{
library("dlstats")
x <- cran_stats(c("dlstats", "emojifont", "rvcheck"), use_cache=TRUE)
head(x)
}
}
\author{
Guangchuang Yu
}
