% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlstats-package.R
\docType{package}
\name{dlstats-package}
\alias{dlstats}
\alias{dlstats-package}
\title{dlstats: Download Stats of R Packages}
\description{
Monthly download stats of 'CRAN' and 'Bioconductor' packages. Download stats of 'CRAN' packages is from the 'RStudio' 'CRAN mirror', see \url{https://cranlogs.r-pkg.org:443}. 'Bioconductor' package download stats is at \url{https://bioconductor.org/packages/stats/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GuangchuangYu/dlstats}
  \item Report bugs at \url{https://github.com/GuangchuangYu/dlstats/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com}

}
\keyword{internal}
