% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bioc_stats.R
\name{plot_bioc_stats}
\alias{plot_bioc_stats}
\title{plot_bioc_stats}
\usage{
plot_bioc_stats(pkg = YGC_bioc_pkg)
}
\arguments{
\item{pkg}{packages}
}
\value{
ggplot object
}
\description{
plot bioconductor download stats
}
\author{
guangchuang yu
}
