% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bioc_stats.R
\name{plot_cran_stats}
\alias{plot_cran_stats}
\title{plot_cran_stats}
\usage{
plot_cran_stats(pkg = YGC_cran_pkg)
}
\arguments{
\item{pkg}{packages}
}
\value{
ggplot object
}
\description{
plot cran download stats
}
\author{
guangchuang yu
}
