% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_stats.R
\name{set_cran_start_year}
\alias{set_cran_start_year}
\title{set_cran_start_year}
\usage{
set_cran_start_year(year)
}
\arguments{
\item{year}{start year}
}
\description{
set query start year for `cran_stats`
}
\author{
Guangchuang Yu
}
