% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc_stats.R
\name{bioc_stats}
\alias{bioc_stats}
\title{bioc_stats}
\usage{
bioc_stats(packages)
}
\arguments{
\item{packages}{packages}
}
\value{
data.frame
}
\description{
monthly download stats of cran package(s)
}
\author{
Guangchuang Yu
}

