% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_stats.R
\name{cran_stats}
\alias{cran_stats}
\title{cran_stats}
\usage{
cran_stats(packages)
}
\arguments{
\item{packages}{packages}
}
\value{
data.frame
}
\description{
monthly download stats of cran package(s)
}
\author{
Guangchuang Yu
}

